bug fixed

+++ src/com/funs/login/dao/LoginDAO.java
@@ -42,7 +42,7 @@ public class LoginDAO extends BaseDAO{
 	}
 	
 	public int queryCountOfPackage(){
-		return this.sqlSessionTemplate.selectOne("com.funs.package.queryCountOfPackage");
+		return (Integer)this.sqlSessionTemplate.selectOne("com.funs.package.queryCountOfPackage");
 	}
 
 }
